requirements.txt
django>=5.2,<6.0
python-telegram-bot[asyncio]>=21.6
langgraph>=0.2.35
mistralai>=0.4.0
psycopg2-binary>=2.9.9
python-docx>=1.1.0
nltk>=3.9.1
django-environ>=0.11.2
pgvector>=0.3.5
pytest>=8.3.0
pytest-django>=4.9.0
pytest-asyncio>=0.23.7
